@@ -1,14 +1,12 @@
 from pas.plugins.authomatic.utils import authomatic_cfg
 from plone.restapi.services import Service
-from typing import Dict
-from typing import List
 
 
 class Get(Service):
     """List available login options for the site."""
 
     @staticmethod
-    def list_plugins() -> List[Dict]:
+    def list_plugins() -> list[dict]:
         """List all configured Authomatic plugins.
 
         :returns: List of login options.
@@ -22,16 +20,14 @@ class Get(Service):
         for provider_id, provider in providers.items():
             entry = provider.get("display", {})
             title = entry.get("title", provider_id)
-            plugins.append(
-                dict(
-                    id=provider_id,
-                    plugin="authomatic",
-                    title=title,
-                )
-            )
+            plugins.append({
+                "id": provider_id,
+                "plugin": "authomatic",
+                "title": title,
+            })
         return plugins
 
-    def reply(self) -> Dict[str, List[Dict]]:
+    def reply(self) -> dict[str, list[dict]]:
         """List login options available for the site.
 
         :returns: Login options information.
