@@ -1,2 +1,8 @@
-from .restapi import RestAPIAdapter  # noQA
-from .zope import ZopeRequestAdapter  # noQA
+from .restapi import RestAPIAdapter
+from .zope import ZopeRequestAdapter
+
+
+__all__ = [
+    "RestAPIAdapter",
+    "ZopeRequestAdapter",
+]
